Move code to bundle handler
